dfx bug fix

+++ frameworks/services/media_dfx/src/dfx_worker.cpp
@@ -71,7 +71,7 @@ static void HandleLoopTask(DfxData *data)
         return;
     }
     int64_t lastReportTime = prefs->GetLong(LAST_REPORT_TIME, 0);
-    int64_t lastMiddleReportTime = prefs->GetInt(LAST_MIDDLE_REPORT_TIME, 0);
+    int64_t lastMiddleReportTime = prefs->GetLong(LAST_MIDDLE_REPORT_TIME, 0);
     DfxManager::GetInstance()->HandleFiveMinuteTask();
     if (MediaFileUtils::UTCTimeSeconds() - lastMiddleReportTime >= SIX_HOUR) {
         MEDIA_INFO_LOG("Report Middle Xml");
